Fix lint errors.
